refactor: badge types
