+++ projects/ion/src/lib/badge/badge.component.spec.ts
@@ -1,8 +1,9 @@
 import { render, screen } from '@testing-library/angular';
-import { IonBadgeComponent, BadgeProps, BadgeType } from './badge.component';
+import { IonBadgeComponent } from './badge.component';
+import { BadgeProps, BadgeType } from '../core/types';
 
 const defaultBadge: BadgeProps = {
-  type: 'primary' as BadgeType,
+  type: 'primary',
 };
 
 const sut = async (customProps: BadgeProps = defaultBadge): Promise<void> => {
