+++ projects/ion/src/lib/badge/badge.component.ts
@@ -1,14 +1,5 @@
 import { Component, Input, OnChanges, OnInit } from '@angular/core';
-import { IconType } from '../icon/icon.component';
-
-export type BadgeType = 'primary' | 'secondary' | 'neutral' | 'negative';
-
-export interface BadgeProps {
-  label?: string;
-  value?: number;
-  icon?: IconType;
-  type: BadgeType;
-}
+import { BadgeType } from '../core/types/badge';
 
 @Component({
   selector: 'ion-badge',
