+++ projects/ion/src/lib/core/types/badge.ts
@@ -1,0 +1,10 @@
+import { IconType } from '../../icon/icon.component';
+
+export type BadgeType = 'primary' | 'secondary' | 'neutral' | 'negative';
+
+export interface BadgeProps {
+  label?: string;
+  value?: number;
+  icon?: IconType;
+  type: BadgeType;
+}
